Rename Stacked plot item component
