+++ src/plugins/plot/vue/stackedPlot/StackedPlotItem.vue
